feat: Add PersonEntity class and its properties
Add a new file `PersonEntity.java` to the `de.ithoc.training.addressbook.repository` package. The file contains the following properties:

- `id` (UUID)
- `firstName` (String)
- `lastName` (String)
- `addresses` (List<AddressEntity>)

The `PersonEntity` class is annotated with `@Entity` and has a no-args constructor. The `id` property is annotated with `@Id` and `@GeneratedValue` to generate a UUID value. The `addresses` property is annotated with `@OneToMany` and `CascadeType.ALL`.

[Repository: /home/oli/ws/address-book-backend]